// src/api/medicine/models/MedicineInventoryEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Medicine Inventory API
 * Medicine inventory management for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MedicineInventoryEntry
 */
export interface MedicineInventoryEntry {
    /**
     * Unique id of the entry in this medicine inventory
     * @type {string}
     * @memberof MedicineInventoryEntry
     */
    id: string;
    /**
     * Name of medicine in medicine inventory
     * @type {string}
     * @memberof MedicineInventoryEntry
     */
    name?: string;
    /**
     * Unique identifier of the medicine known to Web-In-Cloud system
     * @type {string}
     * @memberof MedicineInventoryEntry
     */
    medicineId: string;
    /**
     * Medicine count in the ambulance medicine inventory. It is a number of packages in the medicine inventory for the given ambulance.
     * @type {number}
     * @memberof MedicineInventoryEntry
     */
    count: number;
}

/**
 * Check if a given object implements the MedicineInventoryEntry interface.
 */
export function instanceOfMedicineInventoryEntry(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "medicineId" in value;
    isInstance = isInstance && "count" in value;

    return isInstance;
}

export function MedicineInventoryEntryFromJSON(json: any): MedicineInventoryEntry {
    return MedicineInventoryEntryFromJSONTyped(json, false);
}

export function MedicineInventoryEntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): MedicineInventoryEntry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'medicineId': json['medicineId'],
        'count': json['count'],
    };
}

export function MedicineInventoryEntryToJSON(value?: MedicineInventoryEntry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'medicineId': value.medicineId,
        'count': value.count,
    };
}